상품 보여주는 함수 구현, 찜목록, 내상점 상품 출력
